# Conflicts:
#
PawTrails/PawTrails.xcworkspace/xcuserdata/iosdeveloper.xcuserdatad/User
InterfaceState.xcuserstate
#
PawTrails/PawTrails.xcworkspace/xcuserdata/iosdeveloper.xcuserdatad/xcde
bugger/Breakpoints_v2.xcbkptlist

+++ PawTrails/PawTrails/View/GenderTableViewController.swift
@@ -39,7 +39,7 @@ class GenderTableViewController: UITableViewController {
 
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        let rowGender = Gender(rawValue: indexPath.row)
+        let rowGender = Gender(rawValue: Int16(indexPath.row))
         cell.textLabel?.text = rowGender?.name
         if selected != nil && rowGender == selected {
             cell.accessoryType = .checkmark
@@ -52,7 +52,7 @@ class GenderTableViewController: UITableViewController {
     // MARK: - Table view delegate
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let newSelected = Gender(rawValue: indexPath.row)
+        let newSelected = Gender(rawValue: Int16(indexPath.row))
         selected = newSelected?.code == selected?.code ? nil : newSelected
         tableView.reloadData()
     }
